Share renderHeading with FAQ
