+++ packages/frontend/src/components/index.ts
@@ -2,7 +2,6 @@ export * from './chart/Chart'
 export * from './Footer'
 export * from './head'
 export * from './header/Header'
-export * from './Heading'
 export * from './Logo'
 export * from './navbar/Navbar'
 export * from './OutLink'
