+++ packages/frontend/src/pages/faq/getProps.ts
@@ -1,20 +1,17 @@
-import { Cheerio, Element, load } from 'cheerio'
-import fsx from 'fs-extra'
-import MarkdownIt from 'markdown-it'
 import path from 'path'
 
 import { Config } from '../../build/config'
 import { getFooterProps, getNavbarProps } from '../../components'
+import { getHtml } from '../markdown/getHtml'
 import { Wrapped } from '../Page'
 import { FaqPageProps } from './FaqPage'
-import { renderHeading } from './renderHeading'
 
 export function getProps(config: Config): Wrapped<FaqPageProps> {
   return {
     props: {
       navbar: getNavbarProps(config, 'faq'),
       title: 'Frequently Asked Questions',
-      content: getHtml(),
+      content: getHtml(path.join(__dirname, 'faq.md')),
       footer: getFooterProps(config),
     },
     wrapper: {
@@ -27,38 +24,4 @@ export function getProps(config: Config): Wrapped<FaqPageProps> {
       },
     },
   }
-}
-
-function getHtml() {
-  const markdown = MarkdownIt({ html: true })
-  const file = fsx.readFileSync(path.join(__dirname, 'faq.md'), 'utf-8')
-  const rendered = markdown.render(file)
-  const $ = load(rendered)
-  $('a').each(function () {
-    const $el = $(this)
-    $el.attr('rel', 'noopener noreferrer')
-    $el.attr('target', '_blank')
-  })
-  $('h1, h2, h3, h4, h5, h6').each(function () {
-    const $el = $(this)
-    const html = renderHeading(
-      parseInt(this.tagName[1]),
-      $el.text(),
-      getId($el),
-    )
-    $el.replaceWith($(html))
-  })
-  return $('body').html() ?? ''
-}
-
-function getId($el: Cheerio<Element>) {
-  return (
-    $el.attr('id') ??
-    $el
-      .text()
-      .toLowerCase()
-      .replace(/[^a-z\d]/g, '-')
-      .replace(/-+/g, '-')
-      .replace(/^-+|-+$/g, '')
-  )
 }
