+++ packages/frontend/src/styles/components/_Faq.scss
@@ -12,12 +12,12 @@
     margin: 16px 0 0 0;
   }
 
-  h1.Heading {
+  h1 {
     font-size: var(--font-xl);
     margin: 16px 0 32px 0;
   }
 
-  h2.Heading {
+  h2 {
     margin: 32px 0;
     font-size: var(--font-l);
   }
